Add tags to CTab.
This PR adds tags to CTab widgets. Adding tags through the user interface is currently not included. A simple sorting function is included, so (if you dont add to many equal tags, because it takes them alphabetically) you can sort by tag.

Upon the tags we will in a future PR add the ability to filter based on tags. Additionally, we'll look into adding tags automatically based on certain events such as bookmarking.