Add custom ordering for widgets
This addition should lay the foundation for user defined ordering of widgets. Aside from questions like whether this ordering should be the default or not, or what name it should be given, there is one important question for the implementation: "what should happen when a user drags a widget (and does it matter whether we have the custom ordering selected)?". Right now, after and element has been dragged, the custom ordering state is updated for both the state used by the grid sorting algorithm as well as the state which can be saved to the local storage. We also currently do not take into account which ordering mode is selected. We  always update the grid ordering to the current state regardless of the selected ordering. This means that if we have a custom ordering A and we have currently selected the grid ordering "date added" (B), then if we drag a widget in the current grid view (B'), B' will overwrite the custom order state A. Whether that is desired is to be determined :).